done with post and get endpoints for users
