+++ app.js
@@ -6,12 +6,11 @@ const logger = require('morgan');
 const dotEnv =  require("dotenv");
 dotEnv.config();
 
-const indexRouter = require('./routes/index');
-const usersRouter = require('./routes/users');
-
 
 const relate = require('./database/relationships');
 relate();
+
+const usersRouter = require('./routes/users');
 
 const app = express();
 
@@ -25,7 +24,6 @@ app.use(express.urlencoded({ extended: false }));
 app.use(cookieParser());
 app.use(express.static(path.join(__dirname, 'public')));
 
-app.use('/', indexRouter);
 app.use('/users', usersRouter);
 
 // catch 404 and forward to error handler
